Implement BlitPlanar2Chunky
It works by using "Image transfer across the plane".
The CPU sends each plane as bitstream associated with a BlitFill command.
The graphics card expands each bit to a pixel (using either forground or background
color, set to 0xFF and 0x00 respectively).
The write mask will be used to write to just one individual bit in each pixel
and protect all others.
This way we only transfer the amount of planar data once and get the conversion
to chunky handled by the graphics engine.