Allow tight pitch for resolutions below 640pixel width
Originally, we constrained the width/pitch of surfaces to exactly the ones
supported by the graphics engine, so we can utilize the Blitter
in all desktop resolutions.

Unfortunately, this means, 8bit 320xY will end up with a pitch of 640,
which is unepexted for some PC ports like Rise Of The Triad or
ADoom, leaving them render either garbage or an image of half the height.

Fix this by allowing 'custom' bitmap width for resolutions lower than 640xY.
This will also mean, we cannot use the blitter in these resolutions, but thats
acceptable. We can assume no one runs his Workbench in this resolution and the
CPU will be fast enough to do any blitting - if necessary at all. Typically these
games just lock the bitmap and memcpy() the frames straight into it.

At the same time, remove the restriction that a bitmap needs to fit into 1MB.
I had assumed that blit operations can't straddle 1MB segments because of the
way the segments and blit coordinates are related. But the specs didn't ask for it
and empirically it Just Works(tm). At the same time, this unlocks for resolutions.