Fix GetCompatibleFormats
GetCompatibleFormats helps P96 to decide if SetMemoryMode needs to be called
for a certain format.

On the Trio32/64 all formats are in the one aperture the card provides, thus SetMemoryMode()
is a no-op there. All format that live inn the little endian aperture are always
'compatible' with each other.

For Trio64+ we emulate a few more color formats by making use of the longword- and word byte
swapping mechanism. Bitmaps of formats that requrie the big endian linear address window are
thus only compatible with the formats of the same byte swap setting.
Formats in the little and big endian windows can always coexist and thus are compatbile with
each other.