Accelerate BlitPlanar2Chunky for 320xY modes
In a previous commit, we made it so that 320xY modes get a 320 byte pitch
(as opposed to 640bytes prior) This pitch is not HW supported by the blitter as
it can't be programmed to a pitch less than 640 width resolution.
BlitPlanar2Chunky (and all other blitter related functions) need to fall back
to software emulation for resolution width width < 640 pixels.

I don't expect people using 320xY for their workbench desktop and thus drawing lines,
blitting and filling are not operations typically performed at this resolution.
Instead, these resolutions are mostly used for games.

There's one special case, though. Alien Breed 3D II RTG is using a blit to transfer
a planar image (the animated menu) from chipmem to the display screen, which
triggered P96 calling into BlitPlanar2Chunky.

We enable HW support for BlitPlanar2Chunky for 320xY modes by pretending the
destination display surface has a pitch of 640 and the involved blits have a width of 320
bytes. What this does is that at  the end of each line the HW blitter actually advances
to the next-next line in the destination surface, skipping every other line.

Thus, we split the big blit into two halfs, one for everyty odd and one for every
even line at half the height each. The second blit starts at X offset 320,
which translates to the second line in the destination surface
(from the blitter's point of view the destination surface looks like
a 640xY/2 surface with two 320xY/2 surfaces next to each other)