Optimize BlitTemplate
Since we need to make use of the scissor anyways, use it to also
get rid of the manual bit-ROL due to template->XOffset.
I.e. We  move the blit a bit left and add the same amount of bits to the blit width;
and use the left scissor to cut away this padding.

Also attempt to minimize always waiting for 16 FIFO slots. This might(?)
help with blits that use less than 16 slots (512 monochrome bits)