Use "general pattern blit with rotation" to implement BlitPattern
The hope is that this will be faster for large pattern blits than feeding
the pattern through the HOST_DATA registers for the whole destination
size of the blit (like BlitTemplate).

Contrary to BlitTemplate I would also hope that the pattern is often the
same for consecutive pattern blits, so a cache was implemented to
minimize setup cost and PCI bus transfers.