Test PCI_RETRY
The idea is that instead of polling FIFO_STATUS all the time,
the card is issuing a pci_retry when the FIFO is full -
until there's a FIFO entry open and the transaction succeeds

Prometheus/Firestorm FW support pci_retry, but I don't know
how long the Zorro  timeout is and how that translates into
BUS_FIFO_WS (if at all)

"RETRY mechanism
===============
The new Fire Storm upgrade supports a simple RETRY mechanism for accesses from
Zorro III to PCI. Due to timing constraints on the Zorro III bus it is advisable
to access known slow devices only with PCI-PCI DMA disabled.
For the software side no changes are needed; in case a PCI device does issue a
RETRY situation, the Prometheus CPLDs will repeat the bus access immediately.
If a RETRY fails within the timeout of Zorro III, a Bus Error will occur on
Zorro III.

Cards known to produce RETRYs are:
- slow gfx cards when accessing onboard BIOS ROM
- PCI-PCI bridges, especially on CFG cycles on the PCI bus behind the bridge
"