[S3trio] implement BlitPattern optimization
The old BlitPattern implementation used the CPU to feed the pattern into the mask of a fill
operation. For large pattern fills, this creates a lot of bus traffic with the repeated bit pattern.

But there's a mechanism on the Trio64, 'PATBLT' that allows to splatter a 8x8 pattern into a fill.
The pattern cannot be offset, thus we need to pre-rotate the pattern on the CPU. We cache it
on the graphics card for reuse.