[s3trio] Reading EDID from Monitor
The S3Trio has support to read EDID data from the Monitor. For that it is using
the I2C based "DDC" protocol.

The driver is bit-banging the I2C protocol over the SDA/SCL lines. The monitor is supposed
to respond at address 0x50 with 128 byte of EDID data.

Once we received the data, we write it out to RAM:monitor_name.edid
You can feed this file directly into P96Prefs to make it aware of your monitor's
capabilities.